+++ src/game/data/animType.ts
@@ -16,6 +16,7 @@ export const AnimType = {
     THROW: 15,
     CROUCH: 16,
     PUNCH_1: 17,
+    FOUND_OBJECT: 20,
     DROWNING: 21,
     RUNNING_JUMP: 25,
     TALK: 28,